Fix ReadUInt64

+++ Renci.SshClient/Renci.SshNet/Common/SshData.cs
@@ -202,7 +202,7 @@ namespace Renci.SshNet.Common
         protected UInt64 ReadUInt64()
         {
             var data = this.ReadBytes(8);
-            return (uint)(data[0] << 56 | data[1] << 48 | data[2] << 40 | data[3] << 32 | data[4] << 24 | data[5] << 16 | data[6] << 8 | data[7]);
+            return ((ulong)data[0] << 56 | (ulong)data[1] << 48 | (ulong)data[2] << 40 | (ulong)data[3] << 32 | (ulong)data[4] << 24 | (ulong)data[5] << 16 | (ulong)data[6] << 8 | data[7]);
         }
 
         /// <summary>
